+++ LocalAIHub_Admin/components/logo-mark.tsx
@@ -1,15 +1,13 @@
 'use client'
 
-import Image from 'next/image'
-
 interface LogoMarkProps {
   className?: string
 }
 
 export function LogoMark({ className }: LogoMarkProps) {
   return (
-    <Image
-      src="/logo.png"
+    <img
+      src="/localaihub-admin/logo.png"
       alt="LocalAIHub"
       width={64}
       height={64}